Rename innermost prompt property from "content" to "value"
This makes it clear that "content" is always a peer to "role", whereas "value" only appears when you're using the longhand content array form which takes { type, value } pairs.

Most existing APIs instead do { type: "image", image: theValue }, { type: "text", text: theValue }, etc. This is not really appropriate for a web API, so we don't end up copying them exactly. But none of the existing APIs use "content" at both the outer and inner levels, so it's probably a good idea to avoid that path.