Add Reilly as the editor, retire Domenic (#160)
With Domenic departing Google he is stepping down from his role as
editor of this specification. We thank him for his service to the web
community in this and all of his other work over the years.

With the Chair's permission I nominate myself to take over the role of
editor for this specification.